+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix: [#2797] Single player scenario scores reset when adding a new scenario.
 - Fix: [#3281] Modifying the news options does not trigger a config file write, potentially losing the changes.
 - Fix: [#3314] Players can remove roads owned by other companies.
+- Fix: [#3315] Players can replace station elements owned by other companies.
 
 25.10 (2025-10-10)
 ------------------------------------------------------------------------